requirements.txt
flask==2.2.3
flask-sqlalchemy==3.0.3
flask-jwt-extended==4.5.3
tensorflow-cpu==2.14.0
numpy==1.24.3
werkzeug==2.2.3
python-dotenv==1.0.0
gunicorn==20.1.0